Hash response headers before body
Because we are hashing a known set of headers in a known order, this helps to
prevent two hashes from looking alike even if the headers/bodies are different.
In addition, a mandatory separator is added between the headers and the body.
This separator is generated randomly per-request, but the same separator is used
in a single consensus check.